% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_NULLpipe.R
\name{\%||\%}
\alias{\%||\%}
\title{A utility operator for handling NULL values}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The value to check.}

\item{b}{The value to return if \code{a} is \code{NULL}.}
}
\value{
\code{a} if it is not \code{NULL}; otherwise \code{b}.
}
\description{
This operator returns the left-hand side if it's not \code{NULL}, otherwise it returns the right-hand side.
}
\examples{
a <- NULL
b <- 5
a \%||\% b  # Returns 5

a <- 10
a \%||\% b  # Returns 10
}
