% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_function.R
\name{ProntoPCR}
\alias{ProntoPCR}
\alias{get_css_path}
\title{Run the ProntoPCR Shiny Application}
\usage{
get_css_path()
}
\arguments{
\item{...}{Additional arguments to pass to the ProntoPCR function.}
}
\description{
This function launches the ProntoPCR Shiny app.
}
